RATS RATS WE ARE THE RATS (Aka Sewer expansion
Adjusted the location of the Rouges den
Added few extra nooks to make exploration more intresting / easier
(Mainly: Moved Rouges den closer to the coast and opened a path direcly to it. / Added an extra quest location in a space that before was just blocked off by rock. Added Two ways to get back into the sewers through the river.
Added an entire blackspider condemned part of the sewers (for those wanting to go out and explore)